feat: add decimal

+++ src/SchrodingerServer.Application.Contracts/Token/ITokenPriceProvider.cs
@@ -4,5 +4,5 @@ namespace SchrodingerServer.Token;
 
 public interface ITokenPriceProvider
 {
-    Task<double> GetPriceByCacheAsync(string symbol);
+    Task<decimal> GetPriceByCacheAsync(string symbol);
 }
